refactor: viewmodel 패키지 분리
