+++ app/src/main/java/woowacourse/shopping/view/products/ProductsListFragment.kt
@@ -8,7 +8,7 @@ import androidx.fragment.app.Fragment
 import woowacourse.shopping.R
 import woowacourse.shopping.databinding.FragmentProductListBinding
 import woowacourse.shopping.view.MainActivity
-import woowacourse.shopping.view.MainViewModel
+import woowacourse.shopping.view.viewmodel.MainViewModel
 import woowacourse.shopping.view.cart.ShoppingCartFragment
 import woowacourse.shopping.view.detail.ProductDetailFragment
 import woowacourse.shopping.view.products.adapter.ProductAdapter
