+++ app/src/main/java/woowacourse/shopping/view/viewmodel/MainViewModelFactory.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.view
+package woowacourse.shopping.view.viewmodel
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
